+++ apps/server/src/main.ts
@@ -23,6 +23,7 @@ async function bootstrap() {
 	// create the NestJS application adapting the legacy  server
 	const app = await NestFactory.create(ServerModule, adapter, {});
 
+	// TODO cleanup /api prefix 
 	// for all NestJS controller routes, prepend ROUTE_PREFIX
 	app.setGlobalPrefix(ROUTE_PRAEFIX);
 
@@ -35,5 +36,4 @@ async function bootstrap() {
 
 	adapter.listen(PORT);
 }
-
 void bootstrap();