+++ apps/server/src/modules/database/index.ts
@@ -1,2 +1,2 @@
-export * from './mongo-database';
+// export * from './mongo-database'; currently not in use
 export * from './mongo-memory-database';