+++ apps/server/src/modules/news/controller/dto/create-news.params.ts
@@ -2,7 +2,7 @@ import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
 import { IsDate, IsOptional, IsString } from 'class-validator';
 import { NewsTargetModel } from '../../entity';
 
-const TARGET_MODEL_VALUES = Object.values(NewsTargetModel);
+const TARGET_MODEL_VALUES = Object.values(NewsTargetModel); // Todo does this work?
 
 /**
  * DTO for creating a news document.