+++ apps/server/src/modules/task/controller/task.controller.spec.ts
@@ -1,5 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { TaskRepo } from '../repo/task.repo';
+import { SubmissionRepo } from '../repo/submission.repo';
 import { TaskUC } from '../uc/task.uc';
 import { TaskController } from './task.controller';
 
@@ -15,6 +16,11 @@ describe('TaskController', () => {
 					provide: TaskRepo,
 					useValue: {},
 				},
+				SubmissionRepo,
+				{
+					provide: SubmissionRepo,
+					useValue: {},
+				},
 			],
 			controllers: [TaskController],
 		}).compile();
@@ -25,4 +31,6 @@ describe('TaskController', () => {
 	it('should be defined', () => {
 		expect(controller).toBeDefined();
 	});
+
+	// TODO: test autentication for each endpoint and method
 });