+++ apps/server/src/modules/task/entity/task.entity.ts
@@ -1,10 +1,15 @@
 import { Entity, ManyToOne, Property } from '@mikro-orm/core';
 import { BaseEntityWithTimestamps } from '@shared/domain';
-import { Course } from './course.entity';
-import { Lesson } from './lesson.entity';
+import { CourseTaskInfo } from './course-task-info.entity';
+import { LessonTaskInfo } from './lesson-task-info.entity';
 
 @Entity({ tableName: 'homeworks' })
 export class Task extends BaseEntityWithTimestamps {
+	constructor(partial: Partial<Task>) {
+		super();
+		Object.assign(this, partial);
+	}
+
 	@Property()
 	name: string;
 
@@ -15,8 +20,8 @@ export class Task extends BaseEntityWithTimestamps {
 	private?: boolean;
 
 	@ManyToOne({ fieldName: 'courseId' })
-	course: Course;
+	course: CourseTaskInfo;
 
 	@ManyToOne({ fieldName: 'lessonId' })
-	lesson?: Lesson;
+	lesson?: LessonTaskInfo;
 }