+++ apps/server/test/modules/task/task.e2e-spec.ts
@@ -5,13 +5,13 @@ import { Request } from 'express';
 import { MikroORM } from '@mikro-orm/core';
 import { ServerModule } from '../../../src/server.module';
 import { JwtAuthGuard } from '../../../src/modules/authentication/guard/jwt-auth.guard';
-import { ICurrentUser } from '../../../src/modules/authentication/interface/jwt-payload';
+import { createCurrentTestUser } from '../../../src/modules/user/utils';
 
 describe('Task Controller (e2e)', () => {
 	let app: INestApplication;
 	let orm: MikroORM;
 
-	beforeEach(async () => {
+	beforeAll(async () => {
 		const moduleFixture: TestingModule = await Test.createTestingModule({
 			imports: [ServerModule],
 		})
@@ -19,13 +19,8 @@ describe('Task Controller (e2e)', () => {
 			.useValue({
 				canActivate(context: ExecutionContext) {
 					const req: Request = context.switchToHttp().getRequest();
-					const user: ICurrentUser = {
-						userId: '0000d224816abba584714c9c',
-						roles: [],
-						schoolId: '5f2987e020834114b8efd6f8',
-						accountId: '0000d225816abba584714c9d',
-					};
-					req.user = user;
+					const { currentUser } = createCurrentTestUser();
+					req.user = currentUser;
 					return true;
 				},
 			})
@@ -36,13 +31,19 @@ describe('Task Controller (e2e)', () => {
 		orm = app.get<MikroORM>(MikroORM);
 	});
 
-	afterEach(async () => {
-		await app.close();
-		await orm.close();
+	beforeEach(async () => {
+		//TODO await em.nativeDelete(CollectionName, {});
 	});
 
-	it('/ (FIND)', async () => {
+	afterAll(async () => {
+		await orm.close();
+		await app.close();
+	});
+
+	it('[FIND] /task/dashboard', async () => {
 		const response = await request(app.getHttpServer()).get('/task/dashboard');
 		expect(response.status === 200);
 	});
+
+	it.todo('more /task/dashboard test missing');
 });